+++ doc/source/whats_new.rst
@@ -37,6 +37,8 @@ Changelog
 
    - Add method for computing cross-spectral density (CSD) from epochs and class for storing CSD data by `Roman Goj`_, `Alex Gramfort`_ and `Denis Engemann`_
 
+   - Add Trellis-plot function and method for visualizing single epochs by `Denis Engemann`_
+
 API
 ~~~